+++ src/RcppExports.cpp
@@ -7,22 +7,23 @@
 using namespace Rcpp;
 
 // calc_sum_squares_latent
-NumericVector calc_sum_squares_latent(arma::sp_mat Y, NumericMatrix X, NumericMatrix W, NumericVector ybar);
-RcppExport SEXP _mvrsquared_calc_sum_squares_latent(SEXP YSEXP, SEXP XSEXP, SEXP WSEXP, SEXP ybarSEXP) {
+NumericVector calc_sum_squares_latent(arma::sp_mat Y, arma::mat X, arma::mat W, arma::vec ybar, int threads);
+RcppExport SEXP _mvrsquared_calc_sum_squares_latent(SEXP YSEXP, SEXP XSEXP, SEXP WSEXP, SEXP ybarSEXP, SEXP threadsSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
     Rcpp::traits::input_parameter< arma::sp_mat >::type Y(YSEXP);
-    Rcpp::traits::input_parameter< NumericMatrix >::type X(XSEXP);
-    Rcpp::traits::input_parameter< NumericMatrix >::type W(WSEXP);
-    Rcpp::traits::input_parameter< NumericVector >::type ybar(ybarSEXP);
-    rcpp_result_gen = Rcpp::wrap(calc_sum_squares_latent(Y, X, W, ybar));
+    Rcpp::traits::input_parameter< arma::mat >::type X(XSEXP);
+    Rcpp::traits::input_parameter< arma::mat >::type W(WSEXP);
+    Rcpp::traits::input_parameter< arma::vec >::type ybar(ybarSEXP);
+    Rcpp::traits::input_parameter< int >::type threads(threadsSEXP);
+    rcpp_result_gen = Rcpp::wrap(calc_sum_squares_latent(Y, X, W, ybar, threads));
     return rcpp_result_gen;
 END_RCPP
 }
 
 static const R_CallMethodDef CallEntries[] = {
-    {"_mvrsquared_calc_sum_squares_latent", (DL_FUNC) &_mvrsquared_calc_sum_squares_latent, 4},
+    {"_mvrsquared_calc_sum_squares_latent", (DL_FUNC) &_mvrsquared_calc_sum_squares_latent, 5},
     {NULL, NULL, 0}
 };