+++ tests/testthat/test-mvrsquared.R
@@ -37,6 +37,19 @@ test_that("Get the right r-squared for single column matrix inputs", {
 
 })
 
+test_that("Multithreading works as expected",{
+
+  r2 <- calc_rsquared(y = y, yhat = yhat, threads = 2)
+
+  expect_equal(round(r2, 3), round(s$r.squared, 3))
+
+  ss <- calc_rsquared(y = y, yhat = yhat, ybar = mean(y), return_ss_only = TRUE)
+
+  expect_equal(length(ss), 2)
+
+  expect_equal(r2, 1 - ss[[1]] / ss[[2]])
+
+})
 
 
 ### fancier stuff ----